+++ accessories/netgear/aiowndr.py
@@ -7,6 +7,7 @@ import sys,os,time,logging,re
 sys.path.append(os.path.join(os.path.dirname(os.path.realpath(__file__)), '../..'))
 import lib.tls as tls
 
+
 logger = tls.get_logger(__file__, logging.DEBUG)
 RENUM=r'"([\d,\.]+)"'  #  regex for float number between "
 rxdct = dict(
@@ -16,7 +17,7 @@ rxdct = dict(
 	rxYestday=re.compile(r'yesterday_down='+RENUM)
 )
 
-async def traffic(session: aiohttp.ClientSession, user = 'admin',pwd=None, host = '192.168.1.1'):
+async def traffic(session: aiohttp.ClientSession, user = 'admin',pwd=None, host = '192.168.1.1', semaphore=None):
 	''' gets traffic.htm page from host, parses page looking for items in rxdct '''
 	#url = 'http://' + host +'/traffic_meter_2nd.htm' 
 	#url = 'http://' + host +'/RST_statistic.htm'
@@ -27,15 +28,18 @@ async def traffic(session: aiohttp.ClientSession, user = 'admin',pwd=None, host 
 		auth = aiohttp.BasicAuth(user, pwd)
 	else:
 		auth = None
-	
-	for i in range(2):
+	if semaphore is None:
+		semaphore = asyncio.Semaphore()
+		logger.info('WNDR no synchronisation with other http requests')
+		
+	for i in range(2):  # try it twice
 		try:
-			async with session.request(method='GET', url=url, auth=auth) as response:
+			async with semaphore, session.get(url=url, auth=auth, timeout=2.0) as response:
 				stuff = await response.text()
 				if len(stuff)>1 and response.status<400:
-					break
+					break  # success
 		except Exception as e:
-			logger.warning('%d HTTPerror : %s' % (i,e))
+			logger.warning('%d wndr HTTPerror : %s' % (i,e))
 			stuff = ''
 	resp={}
 	try:
@@ -48,17 +52,19 @@ async def traffic(session: aiohttp.ClientSession, user = 'admin',pwd=None, host 
 			else:
 				logger.warning('no match with %s in %s for %s' % (rx, len(stuff), itm))
 		if len(resp)==0:
-			logger.warning('wndr nothing found in stuff:%s=>%s' % (stuff,rxdct))
+			logger.warning('wndr nothing found in stuff:%s:=>%s' % (stuff,rxdct))
 	except Exception as er:
-		logger.error('%d httpErr:%s' % (i,er))
+		logger.error('%d http parse:%s' % (i,er))
 	return resp
 
-async def get_traffic(interval=15, host='192.168.1.1', pwd=None):
+async def get_traffic(interval=15, host='192.168.1.1', pwd=None, semaphore=None):
 	""" get traffic from wndr router in Mbytes """
 	await asyncio.sleep(interval)
+	#if semaphore is None:
+	#	semaphore = HueBaseDev.Semaphore
 	async with aiohttp.ClientSession() as session:
-		return await traffic(session, pwd=pwd, host=host)
+		return await traffic(session, pwd=pwd, host=host, semaphore=semaphore)
 
 if __name__ == "__main__":
-	traf = asyncio.run(get_traffic(pwd='harrewar'))
+	traf = asyncio.run(get_traffic(pwd='har'))
 	logger.info('\ntraffic:\n%s' % traf)